pyproject.toml
[tool.poetry]
name = "mlops"
version = "0.1.0"
description = "test, prototype, playgroud for airflow"
authors = ["Yu-Lin Huang <yulinhuang23@gmail.com>"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.10"
apache-airflow = "^2.5.1"
mlflow = "1.29.0"
scikit-learn = "1.1.2"
matplotlib = "3.6.2"
boto3 = "^1.21.21"
sqlalchemy = "1.4.27"
jupyter = "^1.0.0"
protobuf = "~3.20"
minio = "^7.1.13"
s3fs = "^2023.1.0"
jupyterhub = "^3.1.1"
configurable-http-proxy = "^0.3.0"
jupyterlab = "^3.5.3"
jupyter-contrib-nbextensions = "^0.7.0"
jupyter-nbextensions-configurator = "^0.6.1"
notebook = "6.1.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
